requirements.txt
aniso8601==9.0.1
apispec==4.5.0
click==8.0.1
Flask==2.0.1
flask-apispec==0.11.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
webargs==8.0.0
Werkzeug==2.0.1
